Rename and create CLI for pyxcodeproj
The prototype of Python Xcode project builder script, "pyproj.py" is renamed to "pyxcodeproj".  I still don't like the name, but it's not necessary to worry about it in its infancy.

I added minimal CLI to the script.  By default, it searches for python source directory, and outputs the xcode project directory in the hack root (YHROOT).